Implemented Prompts CRUD API routes and front

+++ app/profile/page.tsx
@@ -19,12 +19,21 @@ const MyProfile = () => {
     router.push(`/update-prompt?id=${post._id}`);
   };
   const handleDelete = async (post) => {
-    const res = await fetch(`/api/prompt/${post._id}`, {
-      method: "DELETE",
-    });
-    const data = await res.json();
-    if (data.success) {
-      router.reload();
+    try {
+      const hasConfirmed = confirm(
+        "Are you sure you want to delete this prompt?"
+      );
+
+      if (hasConfirmed) {
+        const res = await fetch(`/api/prompt/${post._id}`, {
+          method: "DELETE",
+        });
+        const filteredPosts = posts.filter((p) => p._id !== post._id);
+        setPosts(filteredPosts);
+        router.refresh();
+      }
+    } catch (error) {
+      console.error("An error occurred while deleting the prompt");
     }
   };
   return (
